+++ tests/test_accumulator_fixed.rs
@@ -10,7 +10,9 @@ use simple_accumulator::SimpleAccumulator;
 #[test]
 fn test_sanity_push_in_fixed_capacity() {
     const CAPACITY: usize = 3;
+    println!("Creating...");
     let mut acc = SimpleAccumulator::with_fixed_capacity::<f64>(&[], CAPACITY, true);
+    println!("{acc:?}");
 
     let data = vec![0.0, 1.1, 2.2, 3.3, 4.4];
     for &v in &data {
@@ -44,6 +46,7 @@ fn test_only_n_recent_values() {
     // Create a SimpleAccumulator for size 10
     const CAPACITY: usize = 10;
     let mut acc = SimpleAccumulator::with_fixed_capacity::<f64>(&[], CAPACITY, true);
+    println!("{acc:?}");
 
     // and push values into it.
     for &v in &data {